Remove the unused, monkey-patched field `task_status` from `Task`s.
In `TaskManager`, there was a monkey-patched field `task_status` used,
essentially a duplicate of the `status` field in the `TaskState` class.

Since it was not used anywhere, it was removed from the codebase.

Additionally, the tests were refactor to use the `unittest` methods
for assertions instead of the plain `assert`.